# car-service/docs/swagger.yaml
basePath: /
definitions:
  handler.CreateProductRequest:
    properties:
      name:
        example: Product Name
        maxLength: 100
        minLength: 2
        type: string
      price_cents:
        example: 12000
        type: integer
    required:
    - name
    type: object
  handler.CreateProductResponse:
    properties:
      item:
        $ref: '#/definitions/handler.ProductItem'
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.GetProductResponse:
    properties:
      item:
        $ref: '#/definitions/handler.ProductItem'
    type: object
  handler.GetProductsResponse:
    properties:
      hasNext:
        type: boolean
      items:
        items:
          $ref: '#/definitions/handler.ProductItem'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  handler.ProductItem:
    properties:
      id:
        type: integer
      name:
        type: string
      price_cents:
        type: integer
    type: object
  handler.UpdateProductRequest:
    properties:
      name:
        example: Product Name
        maxLength: 100
        minLength: 2
        type: string
      price_cents:
        example: 12000
        type: integer
    type: object
  handler.UpdateProductResponse:
    properties:
      item:
        $ref: '#/definitions/handler.ProductItem'
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample car ordering server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Backend Example API
  version: "1.0"
paths:
  /cars:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of products with optional filtering and
        sorting.
      parameters:
      - description: field to sort by (name, price_cents)
        in: query
        name: orderBy
        type: string
      - description: sort direction (asc, desc)
        in: query
        name: sortIn
        type: string
      - description: filter by product name (partial match)
        in: query
        name: name
        type: string
      - description: minimum price in cents
        in: query
        name: minPrice
        type: integer
      - description: maximum price in cents
        in: query
        name: maxPrice
        type: integer
      - description: page number
        in: query
        name: page
        type: integer
      - description: items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetProductsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Creates a new product (Admin only).
      parameters:
      - description: Product payload
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a product
      tags:
      - products
  /cars/{id}:
    get:
      consumes:
      - application/json
      description: Returns a single product.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a product by ID
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: Updates product fields (Admin only).
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product update payload
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UpdateProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - products
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"